+++ src/View/Tab/Item/Dropdown.php
@@ -1,25 +1,11 @@
 <?php
 
-/**
- * Contao Bootstrap
- *
- * @package    contao-bootstrap
- * @subpackage Tab
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2013-2018 netzmacht David Molineus. All rights reserved.
- * @license    LGPL-3.0 https://github.com/contao-bootstrap/tab
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Tab\View\Tab\Item;
 
 use ContaoBootstrap\Tab\View\Tab\ItemList;
 
-/**
- * Class Dropdown
- */
 final class Dropdown extends NavItem implements ItemList
 {
     /**
@@ -27,7 +13,7 @@ final class Dropdown extends NavItem implements ItemList
      *
      * @var array|NavItem[]
      */
-    private $items = [];
+    private array $items = [];
 
     /**
      * {@inheritdoc}
@@ -37,19 +23,13 @@ final class Dropdown extends NavItem implements ItemList
         return $this->items;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function addItem(NavItem $navItem): ItemList
+    public function addItem(NavItem $item): ItemList
     {
-        $this->items[] = $navItem;
+        $this->items[] = $item;
 
         return $this;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function active(): bool
     {
         foreach ($this->items as $item) {