+++ src/View/Tab/Navigation.php
@@ -1,16 +1,5 @@
 <?php
 
-/**
- * Contao Bootstrap
- *
- * @package    contao-bootstrap
- * @subpackage Tab
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2013-2018 netzmacht David Molineus. All rights reserved.
- * @license    LGPL-3.0 https://github.com/contao-bootstrap/tab
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Tab\View\Tab;
@@ -19,17 +8,16 @@ use Contao\StringUtil;
 use ContaoBootstrap\Tab\View\Tab\Item\Dropdown;
 use ContaoBootstrap\Tab\View\Tab\Item\NavItem;
 
-/**
- * Class Navigation
- */
+use function in_array;
+
 final class Navigation implements ItemList
 {
     /**
      * Navigation items.
      *
-     * @var array
+     * @var NavItem[]
      */
-    private $items = [];
+    private array $items = [];
 
     /**
      * Create instance from a serialized definition
@@ -41,13 +29,13 @@ final class Navigation implements ItemList
      */
     public static function fromSerialized(string $definition, string $tabId): self
     {
-        $navigation = new Navigation();
+        $navigation = new self();
         $current    = $navigation;
         $definition = StringUtil::deserialize($definition, true);
         $cssIds     = [];
 
         foreach ($definition as $index => $tab) {
-            if (!$tab['cssId']) {
+            if (! $tab['cssId']) {
                 $tab['cssId']  = StringUtil::standardize($tab['title']);
                 $tab['cssId'] .= '-' . $tabId;
 
@@ -66,17 +54,16 @@ final class Navigation implements ItemList
                     $current = $navigation;
                 }
 
-                $item = NavItem::fromArray($tab);
-                $current->addItem($item);
+                if ($current instanceof ItemList) {
+                    $item = NavItem::fromArray($tab);
+                    $current->addItem($item);
+                }
             }
         }
 
         return $navigation;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function addItem(NavItem $item): ItemList
     {
         $this->items[] = $item;