+++ src/View/Tab/TabRegistry.php
@@ -1,16 +1,5 @@
 <?php
 
-/**
- * Contao Bootstrap
- *
- * @package    contao-bootstrap
- * @subpackage Tab
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2013-2018 netzmacht David Molineus. All rights reserved.
- * @license    LGPL-3.0 https://github.com/contao-bootstrap/tab
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Tab\View\Tab;
@@ -18,39 +7,30 @@ namespace ContaoBootstrap\Tab\View\Tab;
 use Assert\Assertion;
 use Contao\ContentModel;
 
-/**
- * Class TabRegistry
- *
- * @package ContaoBootstrap\Tab\View\Tab
- */
 final class TabRegistry
 {
     /**
      * Cached navigation instances.
      *
-     * @var array|Navigation[]
+     * @var Navigation[]
      */
-    private $navigations = [];
+    private array $navigations = [];
 
     /**
      * Cached navigation iterators.
      *
      * @var NavigationIterator[]
      */
-    private $iterators = [];
+    private array $iterators = [];
 
     /**
      * Get a navigation.
      *
      * @param string $elementId The element id.
-     *
-     * @return Navigation
-     *
-     * @throws \Assert\AssertionFailedException When an invalid element is given.
      */
     public function getNavigation(string $elementId): Navigation
     {
-        if (!isset($this->navigations[$elementId])) {
+        if (! isset($this->navigations[$elementId])) {
             $element = ContentModel::findByPk($elementId);
             Assertion::isInstanceOf($element, ContentModel::class);
             Assertion::eq($element->type, 'bs_tab_start');
@@ -65,14 +45,10 @@ final class TabRegistry
      * Get the iterator.
      *
      * @param string $elementId The element id.
-     *
-     * @return NavigationIterator
-     *
-     * @throws \Assert\AssertionFailedException When an invalid element is given.
      */
     public function getIterator(string $elementId): NavigationIterator
     {
-        if (!isset($this->iterators[$elementId])) {
+        if (! isset($this->iterators[$elementId])) {
             $this->iterators[$elementId] = new NavigationIterator($this->getNavigation($elementId));
         }