+++ CHANGELOG.md
@@ -1,10 +1,38 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+### Changed
+
+- Bump minimum PHP version to 7.4
+- Bump Symfony requirements to ^4.4 or ^5.4
+- Bump Contao requirements to ^4.9 or ^4.13
+- Changed coding standard
+- Rewrite content elements to fragment controller
+
+### Added
+
+ - Fix relations when duplicating content elements ([#9](https://github.com/contao-bootstrap/tab/issues/9))
+
+2.0.5 (2020-01-20)
+------------------
+
+[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.0.4...2.0.5)
+
+### Changed
+
+ - Allow symfony/templating ^5.0
+
+### Fixed
+
+ - Fix indents in backend view ([#7](https://github.com/contao-bootstrap/tab/issues/7))
+
 2.0.4 (2018-12-18)
 ------------------
 
-[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.0.1..2.0.3)
+[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.0.3...2.0.4)
 
 ### Changed
 
@@ -13,23 +41,23 @@ Changelog
 2.0.3 (2018-08-24)
 ------------------
 
-[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.0.1..2.0.3)
+[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.0.1...2.0.3)
 
 
 ### Fixed
 
  - Fix missing composer requirements
  - Rework content element to be compatible with Symfony 4
- 
+
 ### Changed
- 
+
  - Updated readme
 
 
 2.0.1 (2018-08-24)
 ------------------
 
-[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.0.0..2.0.1)
+[Full Changelog](https://github.com/contao-bootstrap/tab/compare/2.0.0...2.0.1)
 
 
 ### Changed